requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0
pyyaml>=6.0
psutil>=5.9.0
supabase>=2.0.0


# Connector (Legacy Deps for ctrader-open-api)
ctrader-open-api>=0.9.0
Twisted>=22.10.0
protobuf==3.20.1
python-dotenv>=1.0.0
service_identity>=21.0.0

# ML (Core)
torch>=2.0.0
stable-baselines3>=2.0.0
hmmlearn>=0.3.0
shimmy>=1.0.0

# Dev
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=1.3.0